requirements.txt
Flask==2.3.3
pdfplumber==0.9.0
python-docx==0.8.11
Werkzeug==2.3.7
google-generativeai==0.3.2
fpdf2==2.7.6
